split wrf_netcdf module into reader and transform modules
