+++ gis4wrf/core/transforms/project_to_gdal_checkerboards.py
@@ -1,3 +1,6 @@
+# GIS4WRF (https://doi.org/10.5281/zenodo.1288569)
+# Copyright (c) 2018 D. Meyer and M. Riechert. Licensed under MIT.
+
 from typing import List
 
 from gis4wrf.core.project import Project
